refactor: abstract watcher
